Increased stress duration from 30m to 60m
We observed DevOps Guru takes around 25min to raise the kubernetes pod CPU utilization insight. When we see the insight, even if the incident is already solved, DG shows the insight as on going instead of closed. We expect to make this scenario coherent. The insight will be on going as the stress is still on going.